test/requirements.txt
PyYAML~=6.0.1
requests
mariadb
